+++ src/armies/utils.ts
@@ -1,80 +1,80 @@
-import {Army, armyStatisticsByLevel} from './data';
+import {ArmyAdjustments, armyStatisticsByLevel} from './data';
 
-export function createDefaultArmy(): Army {
+export function getDefaultArmyAdjustment(): ArmyAdjustments {
     return {
-        level: 1,
-        description: '',
-        type: 'infantry',
-        hp: 0,
-        highSave: 'morale',
-        name: 'Army Name',
-        routeThreshold: 0,
-        currentHp: 0,
-        alignment: 'N',
-        consumption: 0,
-        rarity: 'common',
-        adjustments: {
-            ranged: 0,
-            melee: 0,
-            maneuver: 0,
-            recruitmentDC: 0,
-            morale: 0,
-            ac: 0,
-            scouting: 0,
+        ac: 0,
+        melee: 0,
+        morale: 0,
+        maneuver: 0,
+        scouting: 0,
+        recruitmentDC: 0,
+        ranged: 0,
+        highSave: 'maneuver',
+        ammunition: 0,
+    };
+}
+
+function isRanged(item: Item): boolean {
+    return item.system?.weaponType?.value === 'ranged';
+}
+
+export async function addAttackModifiers(actor: Actor, item: Item, level: number, adjustments: ArmyAdjustments): Promise<void> {
+    if (item.type === 'melee') { // true for both melee and ranged
+        const calculated = calculateArmyAdjustments(actor, level, adjustments);
+        const attackModifier = isRanged(item) ? calculated.ranged : calculated.melee;
+        const calculatedUpdate = {system: {bonus: {value: attackModifier}}};
+        await item.update(calculatedUpdate);
+    }
+}
+
+
+export async function syncAttackModifiers(actor: Actor, level: number, adjustments: ArmyAdjustments): Promise<void> {
+    const calculated = calculateArmyAdjustments(actor, level, adjustments);
+    const updates = actor.items
+        .filter(item => item.type === 'melee')
+        .map(item => {
+            const attackModifier = isRanged(item) ? calculated.ranged : calculated.melee;
+            return {'_id': item.id, 'system.bonus.value': attackModifier};
+        });
+    console.log('onactorupdate', updates);
+    await actor.updateEmbeddedDocuments('Item', updates);
+}
+
+
+export function addArmyStats(actor: Actor, update: Partial<Actor>, level: number, adjustments: ArmyAdjustments): void {
+    const calculated = calculateArmyAdjustments(actor, level, adjustments);
+    const calculatedUpdate = {
+        system: {
+            saves: {
+                reflex: {value: calculated.maneuver},
+                will: {value: calculated.morale},
+                fortitude: {value: calculated.recruitmentDC},
+            },
+            attributes: {
+                ac: {value: calculated.ac},
+                perception: {value: calculated.scouting},
+            },
         },
-        gear: [],
-        tactics: [],
-        conditions: [],
     };
+    foundry.utils.mergeObject(update, calculatedUpdate);
 }
 
-export interface CalculatedArmyStats {
-    ac: number;
-    maneuver: number;
-    morale: number;
-    scouting: number;
-    recruitmentDC: number;
-    meleeModifier: number;
-    rangedModifier: number;
-    maximumTactics: number;
-    shots?: number;
-    initiative: number;
-    consumption: number;
-    hp: number;
-    routeThreshold: number;
-}
-
-export interface CalculatedArmy extends Army {
-    calculated: CalculatedArmyStats;
-}
-
-function calculateArmyStats(army: Army): CalculatedArmyStats {
-    const level = army.level;
-    const highSave = army.highSave;
-    const data = armyStatisticsByLevel.get(level)!;
-    const maneuver = highSave === 'maneuver' ? data.highSave : data.lowSave;
-    const morale = highSave === 'morale' ? data.highSave : data.lowSave;
-    const scouting = data.scouting + army.adjustments.scouting;
+export function calculateArmyAdjustments(actor: Actor, level: number, adjustments: ArmyAdjustments): ArmyAdjustments {
+    const data = armyStatisticsByLevel.get(level) ?? armyStatisticsByLevel.get(1)!;
+    const maneuver = adjustments.highSave === 'maneuver' ? data.highSave : data.lowSave;
+    const morale = adjustments.highSave === 'morale' ? data.highSave : data.lowSave;
+    const increasedAmmunitionCount = actor.itemTypes.action
+        .filter(i => i.name.startsWith('Increased Ammunition'))
+        .length;
     return {
-        ac: data.ac + army.adjustments.ac,
-        maneuver: maneuver + army.adjustments.maneuver,
-        morale: morale + army.adjustments.morale,
-        scouting,
-        recruitmentDC: data.standardDC + army.adjustments.recruitmentDC,
-        meleeModifier: data.attack + army.adjustments.melee,
-        rangedModifier: data.attack + army.adjustments.ranged,
-        maximumTactics: data.maximumTactics,
-        consumption: army.consumption,
-        routeThreshold: army.routeThreshold,
-        shots: army.ranged?.shots,
-        initiative: scouting,
-        hp: army.hp,
-    };
-}
-
-export function calculateArmyData(army: Army): CalculatedArmy {
-    return {
-        ...army,
-        calculated: calculateArmyStats(army),
+        ac: data.ac + adjustments.ac,
+        recruitmentDC: data.standardDC + adjustments.recruitmentDC,
+        melee: data.attack + adjustments.melee,
+        ranged: data.attack + adjustments.ranged,
+        maneuver: maneuver + adjustments.maneuver,
+        scouting: data.scouting + adjustments.scouting,
+        morale: morale + adjustments.morale,
+        highSave: adjustments.highSave,
+        ammunition: 5 + adjustments.ammunition + 2 * increasedAmmunitionCount,
     };
 }