+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* Armies
+
 ### Fixed
 
 * Fix broken ingredient links in Hunt & Gather journal
+* Repair reputation now correctly increases control DC by 2
 
 ## [0.6.0] - 2023-09-29